+++ packages/storybook/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @spear-ai/storybook
 
+## 5.6.0
+
+### Minor Changes
+
+- [#314](https://github.com/spear-ai/ui/pull/314) [`cac8bb5`](https://github.com/spear-ai/ui/commit/cac8bb52ebea4786013e61de7227d931bdd6345e) Thanks [@psirenny](https://github.com/psirenny)! - Created Slider component(s).
+
+### Patch Changes
+
+- Updated dependencies [[`cac8bb5`](https://github.com/spear-ai/ui/commit/cac8bb52ebea4786013e61de7227d931bdd6345e)]:
+  - @spear-ai/ui@5.7.0
+
 ## 5.5.1
 
 ### Patch Changes